deliriummcore/api/urls.py
from django.urls import path, include
from rest_framework import routers
from . import views

app_name = 'api'

router = routers.DefaultRouter()
router.register(r'ticket', views.TicketViewSet)
router.register(r'entrada', views.EntradaViewSet)


urlpatterns = [
    path('', include(router.urls)),
    path('api-auth', include('rest_framework.urls'))
]