requirements.txt
asgiref==3.5.2
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
cffi==1.15.1
channels==3.0.5
constantly==15.1.0
crispy-bootstrap5==0.6
cryptography==38.0.1
daphne==3.0.2
Django==4.1.1
django-crispy-forms==1.14.0
djangorestframework==3.14.0
hyperlink==21.0.0
idna==3.4
imgkit==1.2.2
incremental==21.3.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pyOpenSSL==22.1.0
python-decouple==3.1
pytz==2022.4
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.3
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.4.0
zope.interface==5.5.0
